<refact> : 홈피드 API 분리
- 홈피드에 사용되는 API 코드를 기존에 작성된 api 폴더 내의 index.js에서 postAPI.js로 분리
- 분리 후 테스트를 수행하여 정상적으로 작동하는지 확인함.
Close #199
