+++ src/api/index.js
@@ -144,17 +144,17 @@ const FetchApi = {
     return data;
   },
 
-  async loadFeed(token) {
-    const response = await fetch(`${BASE_URL}/post/feed`, {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${token}`,
-        'Content-type': 'application/json',
-      },
-    });
-    const data = await response.json();
-    return data;
-  },
+  // async loadFeed(token) {
+  //   const response = await fetch(`${BASE_URL}/post/feed`, {
+  //     method: 'GET',
+  //     headers: {
+  //       Authorization: `Bearer ${token}`,
+  //       'Content-type': 'application/json',
+  //     },
+  //   });
+  //   const data = await response.json();
+  //   return data;
+  // },
 
   async registerProduct(reqData, token) {
     const response = await fetch(`${BASE_URL}/product`, {
