+++ src/pages/HomeFeed/HomeFeed.jsx
@@ -7,7 +7,7 @@ import PostCard from '../../components/common/PostCard/PostCard';
 import symbolLogoMini from '../../assets/images/symbol-logo-mini.svg';
 import TabMenu from '../../components/common/TabMenu/TabMenu';
 import Nav from '../../components/Nav/Nav';
-import FetchApi from '../../api';
+import postAPI from '../../api/postAPI';
 import * as S from './StyledHomeFeed';
 
 const HomeFeed = () => {
@@ -18,7 +18,7 @@ const HomeFeed = () => {
 
   useEffect(() => {
     const setHomeFeed = async () => {
-      const data = await FetchApi.loadFeed(user.token);
+      const data = await postAPI.loadFeed(user.token);
       setFeed(data.posts);
       setIsLoading(false);
     };
